Unlist old "Power of positive thinking" post
In b581ccf I mistakenly believed I needed to set `published: true` for
the post to appear. In fact this only controls whether the post is
listed on the "All posts" page. Unlist the post (and update the slug)
since it doesn't quite fit with the tone of the blog today.